requirements.txt
langchain>=0.2.14
langchain-openai>=0.1.22
langchain-community>=0.2.12
langgraph>=0.2.6
chromadb>=0.5.5
sentence-transformers>=3.0.1
pydantic>=2.8.2
python-dotenv>=1.0.1
tiktoken>=0.7.0
rich>=13.7.1

fastapi>=0.112.0
uvicorn>=0.30.6
sse-starlette>=2.1.3
rank-bm25>=0.2.2

streamlit>=1.36.0
